Refactor fcm.py and add loader.py   Move from_json and join maps functions to loader.py   Fix var name typo in set_concept_properties, add_concept, and set_default_concept_properties functions
